refactored event layout
