+++ src/components/pages/EventsPage/EventPage.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import CreateEventForm from "@/components/forms/CreateFormEvent/CreateFormEvent";
 import ViewEvents from "./ViewEvents/ViewEvents";
 import { useNavigate, useParams } from "react-router-dom";
-import ViewSingleEvent from "./ViewEvents/SingleEvent/ViewSingleEvent";
+import SingleEventLayout from "./ViewEvents/SingleEvent/SingleEventLayout";
 
 export default function EventPage() {
   const { eventId } = useParams();
@@ -53,7 +53,7 @@ export default function EventPage() {
         {isViewing == "ViewEvents" && <ViewEvents />}
         {isViewing == "ViewSingleEvent" && (
           //@ts-ignore   will not be null
-          <ViewSingleEvent eventId={eventId} setIsViewing={setIsViewing} />
+          <SingleEventLayout eventId={eventId} setIsViewing={setIsViewing} />
         )}
       </div>
     </div>
