+++ src/components/pages/EventsPage/ViewEvents/EventStyles.module.css
@@ -11,8 +11,8 @@
   min-height: 100vh;
   padding: 3em;
   text-align: left;
-  /* display: grid; */
-  /* grid-template-columns: 3fr 2fr; */
+  display: grid;
+  grid-template-columns: 3fr 10fr;
 }
 
 .suggestionsContainer {
